moving to_redshift to Pandas class
